requirements.txt
flask==3.1.2
ChatterBot==1.0.4
chatterbot-corpus==1.2.2
PyYAML==6.0.3
gunicorn==21.2.0
nltk==3.9.2
SQLAlchemy==1.2.19
python-dateutil==2.7.5
